skip parsing of non strict files
Summary:
Profiling reveals that parsing takes about 30% of time when forcing all modules to be strict in a cross module analysis. This diff introduces a preliminary check on files we are about to parse into AST, and avoid the parse altogether if we know that we won't need the AST.

We don't need the AST unless:
- the file contains `__strict__` or `__static__`
- the file is forced to be strict (stubbed, allowlisted, or force_strict returns true)

Somehow doing this reveal a few more memory leaks at shutdown so I tweaked the shut down clean up a bit to make sure the asan tests still pass.

Reviewed By: carljm

Differential Revision: D29187105

fbshipit-source-id: 80e7ce5a99