implement chklist.extend
Summary: The difference from the regular list_extend is that we need to check the elements of iterables as we extend our checked list, and need to deal with the situation where an extend fails halfway through - when dealing with iterables, we don't know when the end is, so need to come up with semantics to handle a type error. The choice I took was to restore the list into it's pre-extension state, and take care of the resize, etc.

Reviewed By: DinoV

Differential Revision: D29885055

fbshipit-source-id: d7b9ba6c57