Use unordered_set to store string literals in LIR parser
Summary:
Originally, the string literals from the parser were stored in a vector. However, the pointer to these strings may get invalidated if the vector resizes.

To prevent this, we can store the strings in an unordered_set. References to strings in an unordered_set shouldn't get invalidated during insertion of new elements. Also, a set prevents duplicates of the same string.

Differential Revision: D29940611

fbshipit-source-id: 4f6ae978da