Implement a preprocessor
Summary:
Add a preprocessing rewriter in to the strict module compiler. This is used between the analyzer and the rewriter (implemented in python and currently in another codebase) as an intermediate step that passes information.
The general strategy is to attach new decorators to classes with auto slotification information so that later the rewriter can consume that information without needing to look at the analysis result.

Reviewed By: DinoV

Differential Revision: D29617298

fbshipit-source-id: 7d76e6ac83