introduce special case for x: Final[dynamic] = value to use value's inferred type
Summary:
Context: This comes into play with regards to functions being declared as protocols. When a user defines

```
a: Final[MyProtocol] = some_function_alias
```

we would like to be able to use static python's extra knowledge with regards to where that function comes from to emit direct invokes.

Since there's no risk of re-assignment due to the final, using the inferred type should be safe in this instance.

Reviewed By: DinoV

Differential Revision: D30015503

fbshipit-source-id: a2e977d564