support indexing checked_list[c_int] via SEQUENCE_GET
Summary:
This optimizes our implementation of checked lists to use the SEQUENCE_GET opcode as appropriate. This elides type checks and boxing/unboxing, and should lead to better perf.

The basic changes are: We support a new sequence kind, SEQUENCE_CHECKED_LIST. No inexact variant needs to be introduced at this time. We then export a `_PyCheckedList_GetItem` API and use that in the interpreter as appropriate, and make the JIT also use this new sequence kind.

Reviewed By: czardoz

Differential Revision: D30046979

fbshipit-source-id: 31953bebcc