ArgMapping improvements/cleanups
Summary:
This is a little bit of improvement around how we use the ArgMapping class.  The idea is to push more of the binding information into this class and have it less scattered around the code base.  So we lose the random checks for `if self.has_vararg or self.has_kwarg:` and those end up in `can_call_statically` which is just a renamed `can_call_self`.

The code generation for all of the calls also ends up in a single `emit` method rather than having multiple callers doing the emitter...  This needs an extra flag to deal with the static method call case where we need `self` to be present in an `INVOKE_METHOD` even though it's not present in the signature.

It also fixes up the type of `args` which really is Optional as built-in function calls it with a super() call which Pyre doesn't seem to be catching.

Reviewed By: sinancepel

Differential Revision: D29951582

fbshipit-source-id: ffcaae1c66