Add TP_ALLOC opcode
Summary: Adds a new `TP_ALLOC` opcode.  This will get tested in a couple of diffs when we actually start generating it.  The opcode simply calls `tp_alloc` on the type.  This will enable us to start constructing types via calling `tp_alloc` and then invoking the `__init__` method directly.

Reviewed By: vladima

Differential Revision: D30005188

fbshipit-source-id: 87a23895e7