generate TP_ALLOC opcodes for calling types
Summary:
This adds basic support for using the `TP_ALLOC` instruction when constructing instances of types.  When we know we have no `__new__` (which really means we have `object.__new__` defined we'll generate a `TP_ALLOC` instruction instead of invoking `__new__`.  If we have a `__new__` function then we'll generate a call to that to produce the instance.

If we use `TP_ALLOC` and we have an `__init__` method then we'll generate a call to the `__init__` method.

Because we're just invoking these methods using our normal method calling methodology this also means we now support calling constructors with primitive values.

In the unlikely event that we have both a custom `__new__` and `__init__` then we just fall back to normal dynamic dispatch - this is extremely unlikely though as usually you pick one.

There's also a new `BuiltinNewFunction` which understands that the return type is going to be the type of `cls`.  To support this the call to the `ArgMapping` on callables is factored into a `map_call` function which can be overridden and returns the `ArgMapping` and the return type.

Reviewed By: czardoz

Differential Revision: D30005185

fbshipit-source-id: cf167182de