CheckedDict __new__/__init__ support
Summary:
Adds some specialization for `CheckedDict.__init__` and `CheckedDict.__new__`.  CheckedDict is a little special in that it has a tp_new which does interesting things (it untracks the object and performs some specialization) but completely ignores the arguments.  `CheckedDict.__init__` then performs the actual initialization of the object like other mutable objects.

Therefore we need to invoke `CheckedDict.__new__` but we can drop all of the arguments, bypassing the limitation of being able to call a type with `__new__` and `__init__` by not duplicating the evaluation of the arguments.

Reviewed By: czardoz

Differential Revision: D30005182

fbshipit-source-id: 1593eddb83