support basic type inference in declaration visitor when something is annotated as `Final[dynamic]`
Summary: The basic problem this solves is that there exist use cases of aliasing functions of the form `f = C.some_static_method` in practice, and currently, the decl visitor drops all type information when dealing with these. In order to be able to generate efficient code for this case, we make a change consistent with what we do in the type binder to use the inferred type for Final[dynamic] (which is sound due to the final-ness of the annotation).

Reviewed By: DinoV

Differential Revision: D30070490

fbshipit-source-id: 43815400c8