Don't update object members, we can't invoke them
Summary:
We don't ever want to generate things like `INVOKE_METHOD` against object, as that would require us to init v-tables for every single object in the system.  So it was a mistake to populate objects members.

This reverts that part of the change, and adds a test to ensure we won't do the `INVOKE_METHOD` on a super call.

Reviewed By: czardoz

Differential Revision: D30114906

fbshipit-source-id: 52aae4708c