Use map of C helper LIR text instead of text files
Summary:
It seems like when we create a runtime package, the python binary cannot find the LIR text files used for inlining (i.e. the files in `Jit/lir/c_helper_translations`). Instead of parsing the text files, we can store the text in a map found in `Jit/lir/c_helper_translations.cpp` and use this map during inlining.

`kCHelperMapping` is a map that stores the function address as a key and the LIR text of that function as the value. During inlining, if the callee address exists in `kCHelperMapping`, we can parse the LIR text and attempt to inline the callee.

Differential Revision: D30027914

fbshipit-source-id: a1fafcf126