replace checked dict opt-out with a checked dict opt-in
Summary: Context: Initially, checked dicts were intended to be the default, and we had a flag to opt people out of chkdicts. However, this ended up causing issues, and we instead went the direction of inferring checked dicts from annotations. This leaves us wanting for a mechanism to opt-in larger parts of the code base to checked dicts. This diff provides a compiler flag to opt-in an entire module to checked dicts, while removing the nonchecked_dicts opt-out compiler flag.

Reviewed By: DinoV

Differential Revision: D30080373

fbshipit-source-id: 2feec3576d