add literal support for CheckedLists
Summary:
This diff adds the ability to create literal checked lists. For now, we only support the `a: CheckedList[int] = [1, 2, 3]`. We support unpacking other checked lists as well during construction.

For the implementation in the interpreter, chose a strategy identical to how we construct checked dicts - we create a single list, and extend it for the sub-lists.

Still to do, will be addressed in future diffs:

1. List comprehension support.
2. JIT'ing BUILD_CHECKED_LIST.
3. Opt-in for checked lists in a given module.

Reviewed By: DinoV

Differential Revision: D30109851

fbshipit-source-id: 250252ccf4