STORE_LOCAL/FAST_LOCAL in global scope should allocate locals
Summary: This is identical to the `LOAD_FAST`/`STORE_FAST` issue we were seeing at the global scope, but it happens with `LOAD_LOCAL`/`STORE_LOCAL` too.  Apparently something in the world has been perturbed and we're now failing a test where we weren't before, despite the fact that we're trashing the stack.

Reviewed By: sinancepel

Differential Revision: D30150601

fbshipit-source-id: 4791a347b1