Allow super() optimization in strict modules
Summary:
Currently we don't get the `super()` optimization in strict modules due to the strict module rewriter.  It will introduce a `super = <fixed-builtins>["super"]` at the start of the module.  This will cause `super` to be a module local instead of an implicit global, and so we give up on the optimization.

That's not really necessary though - if `super` has actually been redefined then we'll fallback to invoking the actual `super` class, or we'll get the optimization in the very likely case that it is still the real `super`.

So this causes us to be a little bit more permissive on when we allow the optimization.

Reviewed By: vladima

Differential Revision: D30144994

fbshipit-source-id: ed8f7c1557