introduce opt-in for checked lists
Summary:
This diff introduces a way for modules to opt in to using checked lists by default. This means that list literals will now be inferred as checked lists, unless explicitly typed as a List[T].

One question: If you look at the latest introduced test, you'll notice that we error when someone tries to assign a checked list to a regular list (this is consistent with checked dicts). Is this what we want longer term, or should we move to considering the LHS context and emitting regular lists/dicts when a user wants one?

Reviewed By: czardoz

Differential Revision: D30136547

fbshipit-source-id: 03e9cc9a8d