add support for directly invoking property gets in final classes
Summary:
This change makes us directly use the `fget` functions of properties in appropriate cases - the property object already places things in the right place, so the work is to extend `classloader_get_member` to deal with property gets correctly, add INVOKE_FUNCTION to the bytecode when appropriate, and add tests.

Next up, we'll add support for method invokes when a class is not final.

Reviewed By: czardoz

Differential Revision: D30261615

fbshipit-source-id: b83dfce373