separate builtins scope
Summary:
The current implementation puts the buildin scope as the module scope. This lead to incorrectness when builtins are aliased and then deleted, where the builtin name should be recovered but was not.
This diff make scope stacks have a separate builtin scope that is also considered global, and make module scope a separate thing. This also means we no longer have to copy builtin scopes since they are not modified at all.

Reviewed By: DinoV

Differential Revision: D30047392

fbshipit-source-id: 62b7ae8e6a