Add strict modules code generator
Summary:
Add `StrictCodeGenerator`. This new generator is going to replace the strict module rewriter to handle all rewrites needed on a strict module. For now, the code generator handles freezing types.
`StaticCodeGenerator` is now a subclass of `StrictCodeGenerator` and will inherit the type freezing behavior.

As part of this, it is also expected that the indicator decorators created by the strict ast preprocessor will be visible to the strict and static code generators. Thus I also added symbol table entries to decorators such as `<mutable>`

Also added a global guard that disables the strict code generator feature by default so that this does not break existing behavior in user code. Once we migrated away from the python implementation of strict modules, this guard can be put away

Reviewed By: sinancepel

Differential Revision: D30122258

fbshipit-source-id: 438624c3e5