Move over definite assignment tests
Summary:
Move over another set of tests that's related to whether names are defined in a strict module.

While moving this over I discovered that the current way of handling finally blocks by analyzing them in a destructor does not work if the finally blocks raise an exception. Changed that part to use a try catch phrase instead.

Reviewed By: sinancepel

Differential Revision: D30296323

fbshipit-source-id: 853a3901f5