Move over class attr conflict checker
Summary:
The class attribute conflict checker is the only checker that's still written in python for strict modules. It checks that a class in a strict module does not declare fields both in class fields and in instance fields, as that will cause a conflict during slotification.
This check is now conducted after the strictness check and before the rewriter, on the original AST of the source.

Reviewed By: sinancepel

Differential Revision: D30296692

fbshipit-source-id: b3bd457072