Revert "Revert "[static] fix chained comparisons""
Summary:
This is a re-do of a reverted change (D29449067 (https://github.com/facebookincubator/cinder/commit/30b705e71334894a65d30e52fc1d50f0e1632060)) that fixes chained comparisons.
The original diff had to be reverted because we saw segfaults in prod.

The segfaults were due to an additional bug which the original change
did not uncover, where we were typing `not i` as a `BOOL_TYPE`, even
if `i` is a primitive. That's wrong, it should be typed as a `CBOOL_TYPE`.

This reverts commit 05ef2a6e1ecd941d899220f64f1748d9e7729ac8.

Reviewed By: sinancepel

Differential Revision: D30024865

fbshipit-source-id: 881da6deb0