support `INVOKE_METHOD` for non-final properties
Summary:
This diff augments our property support to start emitting type descriptors of the form `("module", "Class", ("prop_name", "fget"))` and does all the vtable work to initialize, detect and handle these properly. The end result is that we are able to `INVOKE_METHOD` on these descriptors.

Some general notes: We rely on the format of tuple of two elements being treated specially. We don't support get descriptor overrides of properties in non-static cases, as demonstrated in a test case.

Moral successor of D30371352

Reviewed By: DinoV

Differential Revision: D30405417

fbshipit-source-id: c4ade60bd6