add direct invoke support for class methods
Summary:
We take a similar high level approach to static methods, where we store the classmethod object in
the vtable.

The important idea is that we INVOKE_METHOD iff a class method is called from an instance, as we don't support calls from `Type[C]` and fall back to regular interpreter logic at the moment.
We invoke the underlying callable with a special entry point, where we convert the args to the function so that the self argument's type is passed when we INVOKE_METHOD. Because we only INVOKE_METHOD if a class method is called from an instance, this is fine (tm). Since you INVOKE_FUNCTION on a class method iff self is the class and not an instance, things work out.

Reviewed By: czardoz

Differential Revision: D30525007

fbshipit-source-id: 2b88f02ea5