Annotated assignments at class level with assignment shouldn't produce a slot
Summary:
Currently if we have an annotated assign at the class level with a value we only don't produce a slot if it's a `ClassVar`.  This just means that the user gets an error at runtime when we add the slot.

This makes it so we don't produce the slot whenever we have the annotated assign.

Reviewed By: wmeehan

Differential Revision: D30528863

fbshipit-source-id: 4956251817