Shadow frame pyperf tweaks
Summary:
We were initializing `__strobe_PyThreadState_shadow_frame` to -1 and
conditionally setting it to the real offset at runtime to tell Strobelight if
shadow frame mode was enabled. It reads offsets from the mmaped binary, not
from the in-memory process, so it was never seeing this change. We don't need
this distinction anyway: as long as PyPerf supports shadow frames correctly, it
should do the right thing whether they're enabled globally or not.

Also expose a few constants that Strobelight will need to fully understand
shadow frames, and add `PYSF_PYCODE` as a pointer kind since we expect to need
it for the inliner.

Reviewed By: mpage

Differential Revision: D30591011

fbshipit-source-id: 2534c52f62