Simplify CodeGenerator.visitAugAssign
Summary:
Previously, the compiler used special `Delegator` classes to visit the left-hand side of `ast.AugAssign` statements. The delegators created a fake statement to visit twice, passing a special `mode` argument to distinguish between the first and second calls.

Change this convention to address the three possible left-hand sides individually. This will make it easier to handle each of the cases differently in the static compiler.

Reviewed By: DinoV

Differential Revision: D30587121

fbshipit-source-id: 0b9d4e27a8