fix extra self emission in dynamic calls of classmethods
Summary:
The implementation of classmethod invokes visits the self argument regardless of being called from a class and binds arguments starting from the first parameter of the callable. This contained a bug when dealing with non-statically callable classmethods: It contained an extra visit to self that the super call duplicated, as on the regular python flow, the bound class method would have cls filled in already, and wouldn't need it duplicated.

Change things up so that we only emit the duplicate cls/self parameter when invoking.

Reviewed By: DinoV

Differential Revision: D30697840

fbshipit-source-id: e762719d21