Fix shadowcode handling of changing cached descriptor kind
Summary:
We weren't handling the case where a cached descriptor changed its
"kind" to a data descriptor. The descriptors `__get__` method should
be invoked unconditionally on attribute lookup when that happens.

Unfortunately, there's no way to handle this passively without being
able to watch the descriptor itself for modifications (since it could
be converted to a data descriptor by changing its type at runtime).
We're forced to check whether or not the descriptor is a data descriptor
on each attribute load.

Reviewed By: DinoV

Differential Revision: D30685803

fbshipit-source-id: 3c2add98ad