Update kSymbolMapping to inline more C helper functions
Summary:
`kSymbolMapping` is used to resolve symbols that appear in the translated LIR (e.g. names of functions that are called within the translated function).  To support the parsing of the C helpers, we need more symbols in `kSymbolMapping`.

The newest LIR translations include `JITRT_BoxI32`, `JITRT_BoxU32`, `JITRT_BoxI64`, `JITRT_BoxU64`, `JITRT_UnboxU64`, and `JITRT_UnboxI64`. We add some `PyLong_.*` functions because they are called within the box and unbox functions.

Differential Revision: D30576441

fbshipit-source-id: 770e3604bd