Improve how we handle type modification
Summary:
This cleans up our existing patching functionality.  Currently we try and track changes as they come in, and when we need to figure out the static return type for a given overload we grovel through our v-table data structures to come out with an answer.   There also some assumptions that if we get to a static function it's the original one and is the source of truth.

This adds a more memory intensive, but simpler, system where we just preserve the original v-table members in a side dictionary when we are about to patch a type.  Because this should only happen in testing the overhead should be minimal in production.

It ultimately gets rid of a lot of overhead, but will make it easier for us to find the original function elsewhere too.

A bunch of tests get moved to the `patch.py` file as well.

Reviewed By: carljm

Differential Revision: D30685646

fbshipit-source-id: 4ef0c68dde