Fix bug in `CFG::removeUnreachableBlocks()`
Summary:
`BasicBlock`'s destructor wants the block to have no in-edges, but
this can be violated when we have multiple unreachable blocks that are linked
to each other. I think it's worth keeping the assert around, so avoid tripping
it in `removeUnreachableBlocks()` by separately clearing and then deleting the
unreachable blocks.

Reviewed By: tiansihu

Differential Revision: D30815544

fbshipit-source-id: 520adb7f11