Add thunks for patched INVOKE_FUNCTION
Summary:
We currently have a type safety issue where we're not enforcing the return type of patched functions when we perform an `INVOKE_FUNCTION` against them.

This adds a new `_Py_StaticThunk` which will perform the type enforcement for these cases.  When we resolve a function if we don't get the original function we instead create a thunk and return that instead.  The thunk will then delegate to the original function and perform a type check afterwards.

We previously also had a `funcref` object which we would use for providing an indirect pointer for `INVOKE_FUNCTION` to catch the case where the object was patched.  The thunk object now replaces that w/ a field which either points to the original function or to the thunk in the case where the function has been patched.

Reviewed By: carljm

Differential Revision: D30688792

fbshipit-source-id: daedb35294