Enforce return types on properties
Summary:
After `D30688792` we can now enforce return types on functions, this is just some small fixups so that when we patch properties everything works out okay as well.

One part of this is getting the right function name when reporting errors.

The next part is making sure that `vt_original` includes our properties as well, which we use the `get_func_or_prop_method` helper for.

And finally our logic for getting the original function for the return type now uses `vt_original` rather than simply relying upon a static function being the correct static function.  If that exists we'll consult it, otherwise we know the value is good, or we need to search MRO for the base class version.

Reviewed By: sinancepel

Differential Revision: D30709024

fbshipit-source-id: da78703a75