fix classvar slots
Summary:
Prior to this diff, a class with an implicit (i.e. not explicitly annotated with `ClassVar`) class attribute like

```
class C:
    x: int = 3
```

would not generate a runtime instance slot for `x` but would still emit `LOAD_FIELD` for accesses of `x`,
causing a runtime slot resolution failure and segfault. This is because we didn't check the `is_class_slot`
attribute of the slot in `emit_attr`.

After discussion, we agreed to follow the typing PEPs and require an explicit `ClassVar` annotation on all
class variables, even though this means that in static python a class-level `x: int = 3` will be a compile-time
error for now, since slotified instance attrs can't have class-level default values, and it isn't a class attr without
the explicit `ClassVar[...]`. (In the future we could add a new slot-with-default type to restore more compatible
behavior for `x: int = 3` instead of it being a compile-time error.)

The one exception is a class-level-initialized `Final` attr, which is specifically called out in the typing PEPs as
being implicitly a `ClassVar` (since the initialized value can't be overwritten on instances due to the `Final`).
Thus you are not supposed to ever nest `Final` with `ClassVar` (which is nice, since it simplifies "unwrapping"
these type wrapper annotations to not have to handle arbitrary nesting.)

As part of these changes, I also removed our support for double-annotating an instance attribute (the
`_slot_redefs` stuff), so this is now a compile-time error (the second `: int` should be omitted):

```
class C:
    x: int
    def __init__(self):
        self.x: int = 3
```

This is more consistent with the fact that we don't generally allow re-declaration of types, and saves us from
needing to keep around multiple annotations and decide if they are the same or not.

Also moved some classvar and final tests out of `tests.py` and into `fields.py`.

Reviewed By: DinoV

Differential Revision: D30815045

fbshipit-source-id: 5ac3fb3d62