Ensure __module__ is set on generic types
Summary:
This came up with the strictification of `util.after_party`.  Somehow we're ending up with generic dictionary instances w/o going through `chkdict_cls_getitem` - I thought this might be via the optimized but that doesn't seem to be the case, so there's some other way it slips through.  Regardless we should make sure that when we create the type module is properly initialized.

Now we set the name the same way CPython does based upon the original `tp_name` when constructing the generic type.

Reviewed By: czardoz

Differential Revision: D30746803

fbshipit-source-id: 61b621e7a5