report better error locations in classes
Summary:
Add support for `lineno` and `offset` arguments to `type_error` test utility method,
add lineno and offset to all "fields" tests that test for errors, and update code to
provide better error locations where they were previously not ideal. This required
two main changes: making `InitVisitor` a `GenericVisitor` so that it automatically
sets error context for every node it visits (this allows us to report errors on
individual lines within an `__init__` method rather than always reporting them on
the `def __init__` line) and temporarily tracking a mapping of member names to
nodes during construction of a `Class` so that we can report better locations for
per-member errors that are raised in `finish_bind`.

As part of making `InitVisitor` a `GenericVisitor` I also gave `GenericVisitor` a
`module: ModuleTable` attribute that all its subclasses can rely on, which resulted
in a bunch of renames from `self.cur_mod` to `self.module` in `TypeBinder`, which
previously kept its module under the name `cur_mod`.

Reviewed By: DinoV

Differential Revision: D30854917

fbshipit-source-id: c248dc48ba