Decl visit all of the things!
Summary:
Currently the declaration visitor only visits top-level things, and it just picks out random node types to inspect based upon what we currently support.  This switches to recursively visiting everything, and using the normal visitor pattern for all of the handling of children.

We now have various `declare_` methods which are used for the current object in scope.  For things where we don't want the declaration to be visible this becomes a nop, and we use a `NestedScope` to make child elements entirely a nop.

Reviewed By: sinancepel

Differential Revision: D30848743

fbshipit-source-id: 90be55c4db