fix __patch_enabled__ descriptor
Summary:
In `strictmodule_lookupattro` we have special cases to handle all the data descriptors
defined on the type, and then after checking the instance dict we have an assertion
that whatever descriptor we find at that point must be non-data descriptor. But we were
missing a special-case for the `__patch_enabled__` data descriptor, meaning if it was
accessed on a StrictModule instance we would abort at the "better not be a data
descriptor" assertion. This didn't happen in practice because in the loader we
currently only access the descriptor off the type, never off an instance.

This just adds the missing special case.

Differential Revision: D30903812

fbshipit-source-id: cfb9c2fc3b