Properly align fields for slot access
Summary:
In certain situations involving inheritance, fields might not be
properly aligned in the subclass (see new test for an example). On x86, these
unaligned memory access are occasionally slower but work fine, but there's some
interpreter code that assumes object fields are properly aligned.

I fixed type initialization to align the end of every object so child classes
start off properly aligned, and added an `assert()` to the relevant interpreter
code to defend against similar issues in the future.

Reviewed By: DinoV

Differential Revision: D30883101

fbshipit-source-id: 99ff20c0b4