rename the a.issubclass(b) relationship in the static compiler to mean "is subclass of"
Summary:
When reading `a.issubclass(b)`, I would argue the common interpretation is `a.is_subclass_of(b)`. This diff renames issubclass and changes semantics so that this holds.

There were a couple problems this cleanup addresses:

1. Previously, for basic classes, a.issubclass(b) held if b derived from a, which is the opposite of what one usually means. The generic type relationship was correct. Let's reverse this so that `a.issubclass(b)` holds if a extends b, and deal with the implications in isinstance (the opposite relationship) and can_assign_from.

2. Update the union semantics: What we're looking for here is an "and" relationship.

3. When checking `a <= b | c`, we want `a <= a | b` to hold for all b, but if we require `a <= b` and `a <= a`, this will never hold if a is not a subtype of b, so that needs to be an `any`.

Those three rule changes are the main contribution of this diff, and you can see tests we had to modify to get to a consistent system. One more cosmetic change was to move subclassing- related tests to their own file so that it's easier to follow.

Reviewed By: carljm

Differential Revision: D30939287

fbshipit-source-id: 1b42727df2