Fix a corner case of side effects in Lazy Imports
Summary:
Import time side effects (attributes pointing to child in parent module) should be added (as deferred modules) when creating the deferred object.

This fixes a corner case of lazy imports, where imported module has the same name as a deferred name and it gets overwritten by the side effect.

Reviewed By: DinoV

Differential Revision: D30869457

fbshipit-source-id: 09c3a91338