Fix bug in trivial phi removal
Summary:
When removing a trivial phi, we incorrectly asserted that it would
provide the local definition for whatever non-ssa register it was
replacing. That's not always true: another instruction may redefine
the non-ssa register later in the same basic block.

I think it might be worthwhile to factor out trivial phi removal
into a separate pass. Our current implementation remains faithful
to the paper and is efficient, but that makes it inherently more
complex. Having it as a separate pass/step should be simpler and
easier to reason about.

Reviewed By: swtaarrs

Differential Revision: D31012477

fbshipit-source-id: f5d8ff32a8