Preserve assert statements for injecting casts
Summary:
We have support for emiting casts for assert statements when we're running in an optimized build. But the AST optimizer has already removed the assert statements, so they don't show up.

This adds a new flag to the optimizer to force the asserts to stay around.

Our existing test case was passing because we add the cast for the strongly typed return type.  The test is now updated so that we return `object` and the only source of the cast is from the assert.

Reviewed By: sinancepel

Differential Revision: D31032166

fbshipit-source-id: 44edbe21ec