set effect_name_nodes in reverse apply
Summary: When the asserts are enabled in optimized build we run into an issue where `assert not isinstance(x, Foo)` results in a compile time error because we don't flow the names through via the negated effect.  This adds the extra dictionary to `reverse` and now we collect the names.

Reviewed By: sinancepel

Differential Revision: D31032164

fbshipit-source-id: f5c6d58844