Stop crashing on calling builtins.object methods
Summary:
Currently we emit an INVOKE_METHOD when we encounter
a call to `object.__setattr__`:

```
INVOKE_METHOD            4 ((('builtins', 'type', '__setattr__'), 3))
```

This is crashy behavior, because we don't support `PyWrapperDescr_Type`
values in our vtables.

Even if we add those though, we still shouldn't try to INVOKE_METHOD
them, because then, we'd end up in `type_setattro`, bypassing the normal
runtime type-check.

Another way to fix it is by populating the members of `OBJECT_TYPE`,
however, we definitely do not want to create a v-table for `object`,
because that'll force us to create a v-table for every type.

Instead, fixed it by allowing us to mark `BuiltinMethodDescriptor`s as
eligible for dynamic dispatch. That way, we treat those method lookups
as dynamic, and never emit `INVOKE_*` on them.

Reviewed By: DinoV

Differential Revision: D30939847

fbshipit-source-id: 2071a81bfc