Support PyCodeObjects on JIT-list
Summary:
Allow `PyCodeObject`s on the JIT-list with the format used by the diff below.

The path part of a filename are stripped (leaving the basename). This is needed as paths are generally too specific to allow a portable JIT-list (see the paste in the test plan of the diff below).

Line number matching is optional and off by default. In practice I've observed what tiansihu predicted some time ago which is line numbers are very unstable.

Wildcards are not currently supported.

Reviewed By: tiansihu

Differential Revision: D26783335

fbshipit-source-id: f39a91d6d9