support coroutine invocations of class methods in the classloader
Summary: Whenever the coroutine flag is set, our vtables uses the `type_vtable_coroutine` entry point. This was always executing a raw vector call, which doesn't work with classmethods, where we need to replace the first argument with the type in the args. The strategy taken was to add a case to the `type_vtable_coroutine` instead of adding a new entry point. This does introduce an extra branch in this path, and I'm not sure what we expect the cost of this to be - I can also see a case for a new entry point, curious what reviewers think.

Reviewed By: czardoz

Differential Revision: D31060337

fbshipit-source-id: 11590beff2