Don't use dict keys as guards for split dict lookups in shadowcode
Summary:
The split dictionary layout is an optimization used by CPython to reduce
the memory overhead of instance dictionaries. A dictionary with a split layout
will have its keys and values stored separately, allowing the keys to be
shared between multiple dictionaries subject to a few constraints (keys must
be added in the same order, no deletions can occur, no resizing). Dictionaries
with a split layout are converted to a combined layout (keys and values are commingled)
when any of the constraints are violated.

Instance dictionaries start out with a split layout and typically remain that
way throughout their lifetime. Attributes are typically added unconditionally
in `__init__` and attribute deletions are rare. As a result, the keys object is
typically shared by the instance dictionaries of instances of the same type.

Shadowcode uses this fact to optimize attribute lookups for instances whose
instance dictionaries use a split layout. The first time an attribute is
retrieved, shadowcode will look up the attribute in the dict keys object stored
on the instance's type and cache the resulting offset of the attribute in
the values array (if present). Subsequent lookups can index directly into the
values array using the cached offset, provided that the instance dictionary
uses the same keys object as when the cache entry was initialized.

Previously, shadowcode stored the keys object in the cache entry. It would only
perform a lookup using the cached offset if the instance dictionary's keys object
matched the keys object in the cache entry. This check isn't necessary. For a given
split dict cache entry, its keys object can differ from the keys object of an
instance dictionary of an instance of the same type under the following scenarios:

1. The keys object on the type changes after the cache was initialized and the instance
   was created after the keys changed. In this case the cache entry would have a pointer
   to the old keys object from the type, while the instance would have a pointer to the
   new keys object. This cannot happen; cache entries are invalidated when the keys object
   on a type changes.
2. The instance dictionary had attributes added in an order that diverged from the
   cached keys object or had attributes removed. CPython's dictionary implementation will
   automatically convert the instance dictionary into a combined dictionary, making it
   ineligible for the split dict fast path.
3. The instance dictionary was reassigned using a split dictionary that uses a different
   keys object. We can detect this scenario and convert the dictionary to use a combined
   layout, reducing to case (2).
4. The instance had its type changed (i.e. someone assigned to `__class__`). We can
   detect this scenario and convert the dictionary to use a combined layout, again
   reducing to case (2).

Eliminating the keys check will be a marginal win on its own, but more
importantly it frees up 8 bytes in `_PyShadow_InstanceAttrEntry`. We reuse
this space to optimize method lookup on types, which is a more significant
win.

Reviewed By: DinoV

Differential Revision: D30968830

fbshipit-source-id: a86feae143